Code Review - Xiaoyan Qin
Your codes are neat and have good readability. I fixed a small bug in index.html. There are duplicate meta elements.
One suggestion for the website is that you can change the instruction button to a pop-up message and place the button to a conspicuous area. The question mark as the button is not clear enough for the user to find.
Another suggestion is that you can adjust the background and foreground colors (like button). They do not have a sufficient contrast ratio to meet the WCAG accessibility requirement.
I like your naming "MoneyTing". The idea of the website is really cool. The website has a lot of small and good features, like change user avator/password. 
Great work!